+++ examples/data-sources/aiven_organization/data-source.tf
@@ -1,0 +1,3 @@
+data "aiven_organization" "organization1" {
+  name = "<ORGANIZATION_NAME>"
+}