+++ examples/data-sources/aiven_organizational_unit/data-source.tf
@@ -1,0 +1,4 @@
+data "aiven_organizational_unit" "organizational_unit1" {
+  name      = "<ORGANIZATIONAL_UNIT_NAME>"
+  parent_id = "<ORGANIZATION_ID>"
+}