+++ examples/resources/aiven_organization/resource.tf
@@ -1,0 +1,3 @@
+resource "aiven_organization" "organization1" {
+  name = "<ORGANIZATION_NAME>"
+}